requirements.txt
# Core dependencies
streamlit>=1.28.0
streamlit-mermaid>=0.1.1
pyyaml>=6.0.1
typing>=3.7.4

# PDF and Image Processing
pdf2image>=1.16.3
pillow>=9.5.0

# OpenAI Integration
openai>=1.3.0

# Parsing and transformation
dataclasses>=0.8; python_version < '3.7'

# Optional: For development and testing
pytest>=7.3.1
mypy>=1.3.0

# System compatibility
setuptools>=67.7.2
wheel>=0.40.0